requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
spacy>=3.6.0
gensim>=4.3.0
transformers>=4.30.0
torch>=2.0.0
rapidfuzz>=3.0.0
matplotlib>=3.7.0
reportlab>=4.0.0
plotly>=5.15.0
PyPDF2>=3.0.0
pdfplumber>=0.9.0
python-docx>=0.8.11